:recycle:[refactor]: MOOD-15 <responseDTO 생성자추가>
