+++ src/main/java/com/moodmate/moodmatebe/domain/chat/dto/response/ChatResponseDto.java
@@ -1,14 +1,18 @@
-package com.moodmate.moodmatebe.domain.chat.dto;
+package com.moodmate.moodmatebe.domain.chat.dto.response;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 
 import java.util.List;
 
 @Getter
-@AllArgsConstructor
 public class ChatResponseDto {
     private ChatUserDto user;
     private ChatPageableDto pageable;
     private List<MessageDto> chatList;
+
+    public ChatResponseDto(ChatUserDto chatUserDto, ChatPageableDto chatPageableDto, List<MessageDto> chatList){
+        this.user = chatUserDto;
+        this.pageable = chatPageableDto;
+        this.chatList = chatList;
+    }
 }
